feat: Writing a bitmap file
I implemented the classes and I made a small test. The result is a white
dot that appears in the middle of the screen.

+++ includes/bitmapInfoHeader.h
@@ -9,8 +9,8 @@ struct BitmapInfoHeader{
 	int32_t headerSize{40};
 	int32_t width{0};
 	int32_t height{0};
-	int32_t planes{1};
-	int32_t bitsPerPixel{24};
+	int16_t planes{1};
+	int16_t bitsPerPixel{24};
 	int32_t compression{0};
 	int32_t dataSize{0};
 	int32_t horizontalResolution{2400};
